[DP]: Fix diagnostic protocol create/destroy bugs
Fixed an issue where duplicate instances of the diagnostic protocol would be created when only 1 should exist.
Fixed an issue when deleting a diagnostic protocol, where we might throw an exception when traversing
the list of protocols due to us deleting objects from within that list while traversing it.
Renamed the DM13 test to be more inclusive of all diagnostic protocol tests.
Added a test for creating and destroying diagnostic protocols.